some minor improvements

+++ examples/RayTracer/renderImage.cpp
@@ -8,47 +8,46 @@
 
 #include "renderImage.hpp"
 
-// note: we are using doubles for calculations
 // todo: better vector naming
 
 struct Camera
 {
-    glm::dvec3 eye = {0.0, 0.0, 0.0};
-    glm::dvec3 center = {0.0, 0.0, -1.0};
-    glm::dvec3 up = {0.0, 1.0, 0.0};
-    double fovy = 90.0; // in degrees
+    glm::vec3 eye = {0.f, 0.f, 0.f};
+    glm::vec3 center = {0.f, 0.f, -1.f};
+    glm::vec3 up = {0.f, 1.f, 0.f};
+    float fovy = 90.f; // in degrees
 };
 
 struct Ray
 {
-    glm::dvec3 origin;
-    glm::dvec3 dir;
+    glm::vec3 origin;
+    glm::vec3 dir;
 };
 
 struct Sphere
 {
-    glm::dvec3 center;
-    double radius;
-    glm::dvec3 color;
+    glm::vec3 center;
+    float radius;
+    glm::vec3 color;
 };
 
 struct Plane
 {
-    glm::dvec3 center;
-    glm::dvec3 normal;
-    glm::dvec3 color;
+    glm::vec3 center;
+    glm::vec3 normal;
+    glm::vec3 color;
 };
 
-Pixel toPixel(const glm::dvec3 v)
+Pixel toPixel(const glm::vec3 v)
 {
     Pixel p;
-    p.r = static_cast<unsigned char>(v.r * 255.0);
-    p.g = static_cast<unsigned char>(v.g * 255.0);
-    p.b = static_cast<unsigned char>(v.b * 255.0);
+    p.r = static_cast<unsigned char>(v.r * 255.f);
+    p.g = static_cast<unsigned char>(v.g * 255.f);
+    p.b = static_cast<unsigned char>(v.b * 255.f);
     return p;
 }
 
-bool intersect(const Ray ray, const Sphere& sphere, double* const distance)
+bool intersect(const Ray ray, const Sphere& sphere, float* const distance)
 {
     const auto L = sphere.center - ray.origin;
     const auto tca = glm::dot(ray.dir, L);
@@ -61,23 +60,23 @@ bool intersect(const Ray ray, const Sphere& sphere, double* const distance)
     const auto thc = glm::sqrt(radius2 - d2);
     const auto t0 = tca - thc;
 
-    if(t0 < 0.0)
+    if(t0 < 0.f)
         return false;
 
     *distance = t0;
     return true;
 }
 
-bool intersect(const Ray ray, const Plane& plane, double* const distance)
+bool intersect(const Ray ray, const Plane& plane, float* const distance)
 {
     const auto denominator = glm::dot(ray.dir, plane.normal);
 
-    if(glm::abs(denominator) < 0.001)
+    if(glm::abs(denominator) < 0.001f)
         return false;
 
     const auto t = glm::dot(plane.center - ray.origin, plane.normal) / denominator;
 
-    if(t < 0.0)
+    if(t < 0.f)
         return false;
 
     *distance = t;
@@ -87,7 +86,7 @@ bool intersect(const Ray ray, const Plane& plane, double* const distance)
 void renderImage1(Pixel* buffer, const glm::ivec2 size, std::atomic_int& progress)
 {
     Camera camera;
-    camera.eye = {-1.0, 1.5, 2.0};
+    camera.eye = {-1.f, 1.5f, 2.f};
 
     // camera coordinate system
     const auto z = glm::normalize(camera.eye - camera.center);
@@ -96,16 +95,16 @@ void renderImage1(Pixel* buffer, const glm::ivec2 size, std::atomic_int& progres
 
     const Sphere spheres[] =
     {
-        {{0.0, 0.0, -2.0}, 1.0, {1.0, 0.5, 0.0}},
-        {{-0.8, 0.0, -2.0}, 1.0, {0.0, 0.0, 1.0}},
-        {{0.8, 0.0, -2.0}, 1.0, {0.0, 1.0, 0.0}},
-        {{-1.0, 1.5, -1.5}, 0.5, {1.0, 1.0, 1.0}}
+        {{0.f, 0.f, -2.f}, 1.f, {1.f, 0.5f, 0.f}},
+        {{-0.8f, 0.f, -2.f}, 1.f, {0.f, 0.f, 1.f}},
+        {{0.8f, 0.f, -2.f}, 1.f, {0.f, 1.f, 0.f}},
+        {{-1.f, 1.5f, -1.5f}, 0.5f, {1.f, 1.f, 1.f}}
     };
 
     Plane planes [] =
     {
-        {{0.0, -10.0, 0.0}, {0.0, 1.0, 0.0}, {1.0, 0.0, 0.0}},
-        {{-50.0, 0.0, 0.0}, {1.0, 0.0, 0.0}, {0.0, 0.0, 1.0}}
+        {{0.f, -10.f, 0.f}, {0.f, 1.f, 0.f}, {1.f, 0.f, 0.f}},
+        {{-50.f, 0.f, 0.f}, {1.f, 0.f, 0.f}, {0.f, 0.f, 1.f}}
     };
 
     for(auto& plane: planes)
@@ -113,8 +112,8 @@ void renderImage1(Pixel* buffer, const glm::ivec2 size, std::atomic_int& progres
         plane.normal = glm::normalize(plane.normal);
     }
 
-    const auto aspectRatio = static_cast<double>(size.x) / size.y;
-    const auto scale = glm::tan(glm::radians(camera.fovy / 2.0));
+    const auto aspectRatio = static_cast<float>(size.x) / size.y;
+    const auto scale = glm::tan(glm::radians(camera.fovy / 2.f));
 
     Ray ray;
     ray.origin = camera.eye;
@@ -123,14 +122,14 @@ void renderImage1(Pixel* buffer, const glm::ivec2 size, std::atomic_int& progres
     {
         for(auto i = 0; i < size.x; ++i)
         {
-            glm::dvec3 rayScreenPos;
-            rayScreenPos.x = (2.0 * (i + 0.5) / size.x - 1.0) * scale * aspectRatio;
-            rayScreenPos.y = (1.0 - 2.0 * (j + 0.5) / size.y) * scale;
-            rayScreenPos.z = -1.0;
+            glm::vec3 rayScreenPos;
+            rayScreenPos.x = (2.f * (i + 0.5f) / size.x - 1.f) * scale * aspectRatio;
+            rayScreenPos.y = (1.f - 2.f * (j + 0.5f) / size.y) * scale;
+            rayScreenPos.z = -1.f;
 
             ray.dir = glm::normalize(x * rayScreenPos.x + y * rayScreenPos.y + z * rayScreenPos.z);
 
-            auto distance = std::numeric_limits<double>::infinity();
+            auto distance = std::numeric_limits<float>::infinity();
 
             // it doesn't scale very well, I know
             // at the end we will triangulate everything
@@ -150,7 +149,7 @@ void renderImage1(Pixel* buffer, const glm::ivec2 size, std::atomic_int& progres
 
             for(const auto& sphere: spheres)
             {
-                double t;
+                float t;
 
                 if(intersect(ray, sphere, &t) && (t < distance))
                 {
@@ -162,7 +161,7 @@ void renderImage1(Pixel* buffer, const glm::ivec2 size, std::atomic_int& progres
 
             for(const auto& plane: planes)
             {
-                double t;
+                float t;
 
                 if(intersect(ray, plane, &t) && (t < distance))
                 {
@@ -215,7 +214,7 @@ void renderImage1(Pixel* buffer, const glm::ivec2 size, std::atomic_int& progres
 
 #include "../src/stb_image.h"
 
-void drawLine(glm::ivec2 start, glm::ivec2 end, const glm::dvec3 color,
+void drawLine(glm::ivec2 start, glm::ivec2 end, const glm::vec3 color,
               Pixel* const buffer, const glm::ivec2 imageSize)
 {
     assert(start.x < imageSize.x && start.x >= 0);
@@ -244,8 +243,8 @@ void drawLine(glm::ivec2 start, glm::ivec2 end, const glm::dvec3 color,
 
     for(auto x = start.x; x <= end.x; ++x)
     {
-        const auto t = static_cast<double>(x - start.x) / (end.x - start.x);
-        const auto y = start.y * (1.0 - t) + end.y * t;
+        const auto t = static_cast<float>(x - start.x) / (end.x - start.x);
+        const auto y = start.y * (1.f - t) + end.y * t;
         glm::ivec2 pixelPos(x, y);
 
         if(steep)
@@ -259,21 +258,21 @@ void drawLine(glm::ivec2 start, glm::ivec2 end, const glm::dvec3 color,
 
 struct Triangle
 {
-    glm::dvec3 points[3];
-    glm::dvec2 texCoords[3];
-    glm::dvec3 normals[3];
+    glm::vec3 points[3];
+    glm::vec2 texCoords[3];
+    glm::vec3 normals[3];
 };
 
-glm::dvec3 barycentric(const glm::dvec3* const points, glm::ivec2 P)
+glm::vec3 barycentric(const glm::vec3* const points, glm::ivec2 P)
 {
-    const auto u = glm::cross(glm::dvec3(points[2][0] - points[0][0], points[1][0] - points[0][0], points[0][0] - P[0]),
-                              glm::dvec3(points[2][1] - points[0][1], points[1][1] - points[0][1], points[0][1] - P[1]));
+    const auto u = glm::cross(glm::vec3(points[2][0] - points[0][0], points[1][0] - points[0][0], points[0][0] - P[0]),
+                              glm::vec3(points[2][1] - points[0][1], points[1][1] - points[0][1], points[0][1] - P[1]));
 
     // do we need it? the code seems to work without it
-    //if(glm::abs(u[2]) < 1.0)
-        //return {-1.0, 1.0, 1.0}; // triangle is degenerate, in this case return smth with negative coordinates
+    //if(glm::abs(u[2]) < 1.f)
+        //return {-1.f, 1.f, 1.f}; // triangle is degenerate, in this case return smth with negative coordinates
 
-    return {1.0 - (u.x + u.y) / u.z, u.y / u.z, u.x/ u.z};
+    return {1.f - (u.x + u.y) / u.z, u.y / u.z, u.x/ u.z};
 }
 
 class Texture
@@ -303,42 +302,42 @@ private:
     unsigned char* data_;
 };
 
-Pixel operator*(Pixel p, glm::dvec3 v)
+Pixel operator*(Pixel p, glm::vec3 v)
 {
     return {static_cast<unsigned char>(p.r * v.r),
             static_cast<unsigned char>(p.g * v.g),
             static_cast<unsigned char>(p.b * v.b)};
 }
 
-void drawTriangle(Triangle t, const glm::dvec3 color, Pixel* const buffer, double* const depthBuffer,
+void drawTriangle(Triangle t, const glm::vec3 color, Pixel* const buffer, float* const depthBuffer,
                   const glm::ivec2 imageSize, const Texture& texDiffuse)
 {
     // back-face culling
     {
         const auto n = glm::normalize(glm::cross(t.points[1] - t.points[0], t.points[2] - t.points[0]));
 
-        if(glm::dot(n, {0.0, 0.0, 1.f}) <= 0.0)
+        if(glm::dot(n, {0.f, 0.f, 1.f}) <= 0.f)
             return;
     }
 
     // convert to screen space
     for(auto& point: t.points)
     {
-        point.y *= -1.0;
-        point = (point + 1.0) / 2.0;
-        point *= glm::dvec3(imageSize, 1.0);
+        point.y *= -1.f;
+        point = (point + 1.f) / 2.f;
+        point *= glm::vec3(imageSize, 1.f);
     }
 
     glm::ivec2 start = imageSize - 1;
-    glm::ivec2 end(0.0);
+    glm::ivec2 end(0);
 
     // find the bbox (clipping)
     for(auto i = 0; i < 3; ++i)
     {
         for(int j = 0; j < 2; ++j)
         {
-            start[j] = glm::max(0, glm::min(start[j], static_cast<int>(t.points[i][j] + 0.5)));
-            end[j] = glm::min(imageSize[j] - 1, glm::max(end[j], static_cast<int>(t.points[i][j] + 0.5)));
+            start[j] = glm::max(0, glm::min(start[j], static_cast<int>(t.points[i][j] + 0.5f)));
+            end[j] = glm::min(imageSize[j] - 1, glm::max(end[j], static_cast<int>(t.points[i][j] + 0.5f)));
         }
     }
 
@@ -348,12 +347,12 @@ void drawTriangle(Triangle t, const glm::dvec3 color, Pixel* const buffer, doubl
         {
             const auto b = barycentric(t.points, {x, y});
 
-            if(b.x < 0.0 || b.y < 0.0 || b.z < 0.0)
+            if(b.x < 0.f || b.y < 0.f || b.z < 0.f)
                 continue;
 
             const auto idx = y * imageSize.x + x;
 
-            auto z = 0.0;
+            auto z = 0.f;
 
             for(auto i = 0; i < 3; ++i)
             {
@@ -364,18 +363,18 @@ void drawTriangle(Triangle t, const glm::dvec3 color, Pixel* const buffer, doubl
             {
                 depthBuffer[idx] = z;
 
-                glm::dvec2 texCoord(0.0);
+                glm::vec2 texCoord(0.f);
 
                 for(auto i = 0; i < 3; ++i)
                 {
                     texCoord += t.texCoords[i] * b[i];
                 }
 
-                // (we have flipped points vertically when converting to screen space)
-                texCoord = (texCoord - 1.0) * -1.0;
+                // (we have flipped the points vertically when converting to screen space)
+                texCoord = (texCoord - 1.f) * -1.f;
 
                 const auto texSize = texDiffuse.getSize();
-                const glm::ivec2 texel = texCoord * glm::dvec2(texSize - 1);
+                const glm::ivec2 texel = texCoord * glm::vec2(texSize - 1);
 
                 buffer[idx] = texDiffuse.getData()[texel.y * texSize.x + texel.x] * color;
             }
@@ -414,7 +413,7 @@ struct Model
 
             if(t == "v" || t == "vn")
             {
-                glm::dvec3 v;
+                glm::vec3 v;
 
                 for(auto i = 0; i < 3; ++i)
                 {
@@ -432,7 +431,7 @@ struct Model
             }
             else if(t == "vt")
             {
-                glm::dvec2 texCoord;
+                glm::vec2 texCoord;
 
                 for(auto i = 0; i < 2; ++i)
                 {
@@ -466,9 +465,9 @@ struct Model
         }
     }
 
-    std::vector<glm::dvec3> points;
-    std::vector<glm::dvec2> texCoords;
-    std::vector<glm::dvec3> normals;
+    std::vector<glm::vec3> points;
+    std::vector<glm::vec2> texCoords;
+    std::vector<glm::vec3> normals;
     std::vector<Face> faces;
 };
 
@@ -476,7 +475,7 @@ void renderImage2(Pixel* const buffer, const glm::ivec2 size, std::atomic_int& p
 {
     const auto bufferSize = size.x * size.y;
 
-    std::vector<double> depthBuffer(bufferSize, std::numeric_limits<double>::infinity());
+    std::vector<float> depthBuffer(bufferSize, std::numeric_limits<float>::infinity());
 
     const Model model("res/african_head.obj");
     const Texture texDiffuse("res/african_head_diffuse.tga");
@@ -489,17 +488,17 @@ void renderImage2(Pixel* const buffer, const glm::ivec2 size, std::atomic_int& p
         {
             t.points[i] = model.points[face.points[i]];
             // remove this after we start using projection transformation
-            t.points[i].z *= -1.0;
+            t.points[i].z *= -1.f;
 
             t.texCoords[i] = model.texCoords[face.texCoords[i]];
             t.normals[i] = model.normals[face.normals[i]];
         }
 
         const auto n = glm::normalize(glm::cross(t.points[1] - t.points[0], t.points[2] - t.points[0]));
-        const glm::dvec3 cameraDir(0.0, 0.0, -1.0);
+        const glm::vec3 cameraDir(0.f, 0.f, -1.f);
         const auto intensity = glm::dot(n, -cameraDir);
 
-        drawTriangle(t, glm::clamp(glm::dvec3(1.0, 1.0, 1.0) * intensity, 0.0, 1.0), buffer, depthBuffer.data(),
+        drawTriangle(t, glm::clamp(glm::vec3(1.f, 1.f, 1.f) * intensity, 0.f, 1.f), buffer, depthBuffer.data(),
                      size, texDiffuse);
     }
